refactor: 상수화된 파일을 분리한다
