+++ src/js/constants/index.js
@@ -1,0 +1,15 @@
+import CLASSNAME from "./classname.js";
+import MESSAGE from "./message.js";
+import SNACKBAR_TEXT from "./snackbarText.js";
+import LOCAL_STORAGE_KEY from "./localStorageKey.js";
+import REGULAR_EXPRESSION from "./regularExpression.js";
+import NUMBER from "./number.js";
+
+export {
+  CLASSNAME,
+  MESSAGE,
+  SNACKBAR_TEXT,
+  LOCAL_STORAGE_KEY,
+  REGULAR_EXPRESSION,
+  NUMBER,
+};
