+++ src/js/constants/localStorageKey.js
@@ -1,0 +1,9 @@
+const LOCAL_STORAGE_KEY = Object.freeze({
+  WATCH_LATER_VIDEO_ITEMS: "watchLaterVideoItems",
+  HISTORY_VIDEO_ITEMS: "historyVideoItems",
+  SAVED_VIDEOS_COUNT: "savedVideosCount",
+  KEYWORD_HISTORY: "keywordHistory",
+  QUERY: "query",
+});
+
+export default LOCAL_STORAGE_KEY;
