+++ src/js/constants/message.js
@@ -1,0 +1,14 @@
+const MESSAGE = Object.freeze({
+  KEYWORD_SUBMITTED: "KEYWORD_SUBMITTED",
+  DATA_LOADED: "DATA_LOADED",
+  SAVE_VIDEO_BUTTON_CLICKED: "SAVE_VIDEO_BUTTON_CLICKED",
+  CANCEL_VIDEO_BUTTON_CLICKED: "CANCEL_VIDEO_BUTTON_CLICKED",
+  SAVED_VIDEOS_COUNT_CHANGED: "SAVED_VIDEOS_COUNT_CHANGED",
+  SAVE_IF_VIDEOS_COUNT_IS_IN_RANGE: "SAVE_IF_VIDEOS_COUNT_IS_IN_RANGE",
+  HIDE_IF_VIDEO_IS_SAVED: "HIDE_IF_VIDEO_IS_SAVED",
+  WATCHED_ICON_CLICKED: "WATCHED_ICON_CLICKED",
+  WATCH_LATER_ICON_CLICKED: "WATCH_LATER_ICON_CLICKED",
+  SAVED_VIDEO_DELETED: "SAVED_VIDEO_DELETED",
+});
+
+export default MESSAGE;
