+++ src/js/constants/number.js
@@ -1,0 +1,9 @@
+const NUMBER = {
+  MAX_SAVED_VIDEOS_COUNT: 100,
+  MAX_KEYWORDS_COUNT: 3,
+  MAX_RESULTS_COUNT: 10,
+  SCROLL_EVENT_THRESHOLD: 0.7,
+  SNACKBAR_SHOW_TIME_IN_MS: 3000,
+};
+
+export default NUMBER;
