+++ src/js/constants/regularExpression.js
@@ -1,0 +1,6 @@
+const REGULAR_EXPRESSION = Object.freeze({
+  EMPTY_STRING: /^$/,
+  NOT_EMPTY_STRING: /.+/,
+});
+
+export default REGULAR_EXPRESSION;
